ai-service/requirements.txt
# Core ML dependencies
torch>=2.0.0
torchvision>=0.15.0

# Image processing
pillow>=9.0.0
opencv-python>=4.8.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# API framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0
